Remove failing test case
I spoke with the people who maintain ListenBrainz in the IRC chat for
the project and this is a result of API limitations on the LB side.
There's nothing that can be done to resolve it from the client side, but
recent commits to prioritise reading the MBID, if any, from metadata
should mean that these edge cases are more difficult to reach.